pyproject.toml
[project]
name = "practice-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "sentence-transformers>=5.4.0",
]
